Refactor for production builds
