+++ backend/api-gateway/Dockerfile
@@ -10,6 +10,6 @@ COPY . .
 
 RUN npm run build
 
-EXPOSE 3000
+EXPOSE 8000
 
-CMD ["npm", "run", "start"]
+CMD ["npm", "run", "start:prod"]
