+++ backend/chat-service/package.json
@@ -26,6 +26,7 @@
     "axios": "^1.7.7",
     "reflect-metadata": "^0.2.0",
     "rxjs": "^7.8.1",
+    "socket.io": "^4.8.0",
     "ws": "^8.18.0",
     "y-websocket": "^2.0.4",
     "yjs": "^13.6.20"
@@ -37,6 +38,7 @@
     "@types/express": "^4.17.17",
     "@types/jest": "^29.5.2",
     "@types/node": "^20.3.1",
+    "@types/socket.io": "^3.0.1",
     "@types/supertest": "^6.0.0",
     "@types/ws": "^8.5.12",
     "@typescript-eslint/eslint-plugin": "^8.0.0",
