+++ backend/user-service/Dockerfile
@@ -8,8 +8,6 @@ RUN npm install
 
 COPY . .
 
-RUN npm run build
-
-EXPOSE 3000
+EXPOSE 8003
 
 CMD ["npm", "run", "start"]
